move ChildUpdatePropagationPermissiveness to common state
